2013_04_16: Great Success!

+++ pes_fitting_programs/EMT_potential/atom_class.f95
@@ -2,9 +2,9 @@ module atom_class
     implicit none
 
     type atom
-        real, dimension(3)    :: r=3  ! position
-        real, dimension(3)    :: v=1  ! velocity
-        real, dimension(3)    :: f=2  ! force
+        real(8), dimension(3)    :: r=3  ! position
+        real(8), dimension(3)    :: v=1  ! velocity
+        real(8), dimension(3)    :: f=2  ! force
     contains
 
     end type atom
